Add bower_components for OnPrem release
This ensures that any further patches/changes that need to be made to the OnPrem (10.6.11) BSI release will use the same bower package versions that were initially deployed.

This also removes the bower install step from Travis - it seems that bower sort of just... ignores the bower_components folder, and fetches everything anew. It's entirely strange, and if that is indeed the case, entirely dumb, but this is one solution to that problem.