Lock bower components versions
>> Dave Lockhart [5:22 PM] 
One other thing to note — for the branch of BSI, `bower_components` is checked in, which is nice because it’ll let you see what changes happen to any dependency. What I’d also suggest before making any changes is to lock the version of every dependency in `bower.json` by removing the `^` in front. That way they won’t auto-upgrade beyond what version is explicitly set there.

[5:23] 
Then you can make explicit choices based on what defects you’re trying to fix about when to actually upgrade one of those dependencies. `^1.0.0` means it’ll auto-upgrade to `1.10` or `1.99999` all the way up to but not including `2.0`.