Use d2l-fetch from bower, to take advantage of version resolution
Under this new method, `.use`ing the various middlewares is done by the consumer of the element, in this case BSI. This means we can get the same versions for everything, and the same middlewares. This also includes bumps to a couple other packages that have also been updated to use d2l-fetch.