DE29004 Update d2l-image-banner-overlay (and d2l-search-widget-ui)
Both of these version bumps contain changes to use siren-parser-import, which allows us to pull in a more-recent version of siren-parser than we were getting when these two were referencing the script on the CDN directly. The root cause of the DE was simply that we ended up with the old version clobbering the new one, which meant we were missing some helper functions that were required in My Courses.